[Decomp][Player] Cohesion split: extract rest methods to PlayerRest.cpp
